init_glide.js created. new? added to event policy
