+++ app/assets/stylesheets/application.scss
@@ -7,8 +7,7 @@
 @import "bootstrap/scss/bootstrap";
 @import "font-awesome-sprockets";
 @import "font-awesome";
-@import "../../../node_modules/@glidejs/glide/src/assets/sass/glide.core";
-@import "../../../node_modules/@glidejs/glide/src/assets/sass/glide.theme";
+
 
 // Your CSS partials
 @import "components/index";
