+++ app/policies/event_policy.rb
@@ -29,4 +29,9 @@ class EventPolicy < ApplicationPolicy
   def dashboard?
     true
   end
+
+  def new?
+    true
+  end
+
 end
